requirements.txt
# ── Web Framework ─────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# ── GraphQL ───────────────────────────────────────────────────
strawberry-graphql[fastapi]==0.253.1

# ── Database / ORM ────────────────────────────────────────────
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet>=3.0.0          # required by SQLAlchemy async engine

# ── Redis ─────────────────────────────────────────────────────
redis[hiredis]==5.2.1

# ── RabbitMQ ──────────────────────────────────────────────────
aio-pika==9.5.3

# ── HTTP Client ───────────────────────────────────────────────
httpx==0.28.1

# ── Auth / JWT ────────────────────────────────────────────
python-jose[cryptography]==3.3.0
bcrypt>=4.0.0

# ── Auth / PASETO ────────────────────────────────────────
pyseto>=1.7.0

# ── Settings / Config ─────────────────────────────────────────
pydantic>=2.7.0,<2.11.0   # strawberry-graphql 0.253.1 không tương thích pydantic>=2.11
pydantic-settings==2.7.0
python-dotenv==1.0.1

# ── Logging ───────────────────────────────────────────────────
loguru==0.7.3
